VIITE-2371 adjust terminated roadways for two track cases
