+++ digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/process/strategy/TrackCalculatorStrategy.scala
@@ -170,11 +170,13 @@ trait TrackCalculatorStrategy {
 
     val leftRoadwayNumber = leftProjectLinks.headOption.map(_.roadwayNumber).getOrElse(NewIdValue)
     val continuousLeftProjectLinks = leftProjectLinks.takeWhile(pl => pl.roadwayNumber == leftRoadwayNumber)
+    val restLeft = leftProjectLinks.drop(continuousLeftProjectLinks.size) ++ restLeftProjectLinks
 
     val rightRoadwayNumber = rightProjectLinks.headOption.map(_.roadwayNumber).getOrElse(NewIdValue)
     val continuousRightProjectLinks = rightProjectLinks.takeWhile(pl => pl.roadwayNumber == rightRoadwayNumber)
+    val restRight = rightProjectLinks.drop(continuousRightProjectLinks.size) ++ restRightProjectLinks
 
-    // Find a calibration point annexed to the projectLink Id
+    //  Find a calibration point annexed to the projectLink Id
     val availableCalibrationPoint = calibrationPoints.get(continuousRightProjectLinks.last.id).orElse(calibrationPoints.get(continuousLeftProjectLinks.last.id))
 
     val startSectionAddress = startAddress.getOrElse(getFixedAddress(continuousLeftProjectLinks.head, continuousRightProjectLinks.head)._1)
@@ -182,16 +184,17 @@ trait TrackCalculatorStrategy {
 
     val (adjustedLeft, adjustedRight) = adjustTwoTracks(continuousRightProjectLinks, continuousLeftProjectLinks, startSectionAddress, estimatedEnd, calibrationPoints)
 
-    //The getFixedAddress method have to be called twice because when we do it the first time we are getting the estimated end measure, that will be used for the calculation of
-    // NEW sections. For example if in one of the sides we have a TRANSFER section it will use the value after recalculate all the existing sections with the original length.
+    //  The getFixedAddress method have to be called twice because when we do it the first time we are getting the estimated end measure, that will be used for the calculation of
+    //  NEW sections. For example if in one of the sides we have a TRANSFER section it will use the value after recalculate all the existing sections with the original length.
     val endSectionAddress = getFixedAddress(adjustedLeft.last, adjustedRight.last, availableCalibrationPoint)._2
 
     TrackCalculatorResult(
       setLastEndAddrMValue(adjustedLeft, endSectionAddress),
       setLastEndAddrMValue(adjustedRight, endSectionAddress),
       startSectionAddress, endSectionAddress,
-      leftProjectLinks.drop(continuousLeftProjectLinks.size) ++ restLeftProjectLinks,
-      rightProjectLinks.drop(continuousRightProjectLinks.size) ++ restRightProjectLinks)
+      restLeft,
+      restRight
+    )
   }
 
   /**
